Moved async code to another module
